requirements.txt
diffusers < 0.5, >= 0.4
transformers
mastodon.py == 1.5.1
ftfy
spacy
pillow
beautifulsoup4 >= 4
requests
accelerate
deepspeed
scipy